Refactor MLT XML conversion with proxies to just-in-time.
This frees the file.output_path for holding the rendered output, prevents
storing proxy XML files in transcodes, and allows for bug fixes to not require
rerunning validate jobs. Now, the MltXmlHelper::substituteProxyFiles() does
not write to the database and a similar new function checkFileReferences()
checks the state of dependencies updating the database as needed for
the validate job.