Prevent reundant publishing.
After clicking Publish, the page did not reflect any change, and refreshing
the page to re-check status would resubmit the file for publishing! Also,
the Publish action was still displayed begging to published again. Now,
the approved status flag is set, and the browser redirected to the file
view page. And, the file view page shows the approved status as "Publish
Pending" without the Publish action link.