Fix rendering more than one job within worker.
The childFiles array would not be updated from one job to the next because it was being supplied to the MltXmlWriter constructor.
Also, this fixes rendering a kdenlivetitle with XML entities used to escape like ampersand as part of the text of the title.
Lastly, it includes some minor logging fixes.